rnode: change id type to size_t

+++ libnopegl/src/pass.c
@@ -521,7 +521,7 @@ int ngli_pass_prepare(struct pass *s)
     struct pipeline_desc *desc = ngli_darray_push(&s->pipeline_descs, NULL);
     if (!desc)
         return NGL_ERROR_MEMORY;
-    ctx->rnode_pos->id = (int)ngli_darray_count(&s->pipeline_descs) - 1;
+    ctx->rnode_pos->id = ngli_darray_count(&s->pipeline_descs) - 1;
 
     memset(desc, 0, sizeof(*desc));
